# Rewrote A* to be faster with fewer allocations.