Rewrote A* to more closely follow widely-adopted implementations, greatly speeding up its runtime (roughly 200x faster) and reducing its memory usage (roughly 300x fewer allocations and size of allocations).