@BorisIvanovic 2019-07-09
Thanks! Here's the benchmark I ran using the following graph:
```
graph = barabasi_albert(1000, 100, 100, seed=15)
@profile (for i=1:1000; a_star(graph, 1, 1000); end)
```

And here are the results I got:
### OLD
```
BenchmarkTools.Trial: 
  memory estimate:  331.34 MiB
  allocs estimate:  8710892
  --------------
  minimum time:     2.251 s (14.14% GC)
  median time:      2.319 s (17.01% GC)
  mean time:        2.320 s (16.80% GC)
  maximum time:     2.392 s (19.10% GC)
  --------------
  samples:          3
  evals/sample:     1
```

### NEW
```
BenchmarkTools.Trial: 
  memory estimate:  293.77 KiB
  allocs estimate:  8764
  --------------
  minimum time:     1.419 ms (0.00% GC)
  median time:      1.664 ms (0.00% GC)
  mean time:        1.824 ms (3.26% GC)
  maximum time:     75.829 ms (97.60% GC)
  --------------
  samples:          2716
  evals/sample:     1
```

So maybe a bit better than what I reported above :P 

As for where we got the improvement, there were two major areas:
1. Array concatenation `new_path = cat(path, E(u, v), dims=1)` (line 26 of astar.jl) is very slow. That actually eats up the majority of the runtime before.
2. Hashing an ever-increasing array `enqueue!(frontier, ...` (lines 28-31 of astar.jl) is very slow. 

Both of these can be seen in the attached ProfileView graph of the old method (the big red pyramid on the left is the concatenation, the towers on the right are priority queue hashing):

![old_astar](https://user-images.githubusercontent.com/8534290/60851621-bc76f100-a1a8-11e9-9842-bd52cad38c03.png)

whereas our new implementation has its runtime cost spread out over many different underlying functions:

![new_astar](https://user-images.githubusercontent.com/8534290/60851658-e4feeb00-a1a8-11e9-9f37-8022e62cab2c.png)

Unfortunately, I can't upload the actual SVG outputs, but you could obtain these by running
```
Profile.clear()
@profile (for i=1:1000; a_star(graph, 1, 1000); end)
```
(just know that for the old method, you'll have to change the `1000` to a `1` otherwise Profile's internal buffer will saturate).