@sbromberger 2019-07-09
I really appreciate your attention to detail in the code and the benchmarks. Once we have tests passing (not your fault; we have randomness in a few functions that causes failures sometimes) I'm happy to merge this. Thank you.