@simonschoelly 2019-07-09
This happened quite fast, so I don't know if there is a point in still doing a code review.

One thing that I noticed is, that the result type will change from `Vector{Edge{eltype(g)}}` to `Vector{edgetype{g}}`. This is the same for `SImpleGraph` and `SimpleDiGraph`, but not necessarily for other graphs. I don't know if the constructor `edgetype(g)(u, v)` works for all graphs (What if there is metadata on an edge?)

Also, when doing benchmarks, one should maybe also do some with non-constant edge weights and a non-constant heuristic.

For profiling, I think there is a command, that can increase the number of samples collected.