@sbromberger 2019-07-09
Anything that gives us a 3 orders of magnitude speedup on SimpleGraphs is a no-brainer to me.

As far as the return value goes, my opinion is that the latter (`edgetype(g)`) is more correct: it's part of the API, and since any graph `g` must have edges of `edgetype(g)`, this function should return structures that are valid edges of `g`. `Edge{eltype(g)}` might not always be a valid edgetype for `g`. I can't think of am example now where this isn't the case, but even if there are no such examples, `edgetype(g)` is a fair - and more appropriate - substitute.

If there is metadata on the edge, it should have reasonable default values. The algorithm does not claim to create or preserve metadata of the returned edges. (Perhaps we should consider a `make_edge_from(u, v, e::Edge)` function, but that's outside the scope of this discussion and it seems pointless to copy metadata into a list of edges that represent a path.)

As far as benchmarks, you're right that more formal benchmarking is nice, but I have not seen that in many PRs; in fact, this is some of the most thorough benchmarking and analysis I've seen for any PR in recent memory and would like to encourage this sort of output.

If there is a specific objection to the code – if you've found a regression – then I'm all ears and we should revisit. But as it stands, we've got a massive improvement in speed and it passes all our current tests.