@simonschoelly 2019-07-09
Apologies, if that came over as some harsh criticism, that was never the intention and the comment on the benchmark should be seen more as some kind of feedback, as it should be clear from the code changes, that this code will run faster.

There are two problems with using `edgetype(g)`. The first is, that depending on the graph we might waste a lot of time. Using `Edge{eltype(g)}` does of course also use twice as much space as necessary, but at least not more, and the taken edges can easily be reconstructed (at long as the graph is not a multigraph at least).

The second one is, that the algorithm uses the constructor `edgetype(g)( u, v)` which does exist for SimpleGraphs but not necesarily for others (and in my opinion also should not necessarily). The following code did work before, but does crash now, so this will be a breaking change:

```
using LightGraphs, SimpleWeightedGraphs

A = Float64[0 1; 1 0]
g = SimpleWeightedGraph(A)

a_star(g, 1, 2)
```
I can create a PR to change that to the old behaviour.