@sbromberger 2019-07-09
MHO is that the problem is not with this code, it is with the fact that `SimpleWeightedEdge` does not have a constructor `SimpleWeightedEdge{T, U}(u::T, v::T, w::U=one(U))`. That's probably the correct fix, and we should require any custom edges to be able to be constructed with two integers.