@sbromberger 2019-07-09
This fixes it for SWG:
```
SimpleWeightedEdge{T,U}(u::T,v::T) where {T<:Integer, U<:Real} = SimpleWeightedEdge(u, v, one(U))
```